requirements.txt
streamlit
google-generativeai
python-dotenv
IPython
langchain
PyPDF2
chromadb
faiss-cpu
langchain_google_genai